hwc_vsync: Fix Realtime priority usage
Commit https://github.com/AOSPA/android_hardware_qcom_display/commit/8490450b6bb73adf1327ffd320ade1cbb89cadd0 was
relying on a mix of PRIORITY_URGENT_DISPLAY and PRIORITY_REALTIME.
That results in a huge nice value, which makes the task seem
important.

However that was missing the real purpose of
the PRIORITY_REALTIME definition.
By summing the values of the two definitions
the Threads controller wasn't able to recognize
ANDROID_PRIORITY_REALTIME which is hardcoded at -21
and apply the important SP_REALTIME scheduling policy
on it.

This commit fixes the flaws by relying on the
PRIORITY_REALTIME definition only.

Change-Id: I569b41c9b31d4c0ea8fd186e1894cc0699b4c7de
Signed-off-by: Alex Naidis <alex.naidis@linux.com>